+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Method to check if drive is connected via a COM-KIT.
 - Property to check if the PDOs are active.
 - Method to get a servo's FSoE master handler state.
+- Method to subscribe to the FSoE errors.
 
 ## [0.8.2] - 2024-07-15
 ### Added